fix chart data code
